Puffle statistics & running away
Puffles will now lose health (at a rate of about 1% every half hour). Puffles will not lose health when the player is in the igloo with them but will lose health faster when they're going for a walk. Appropriate postcards are sent out when puffle is hungry and when it has run away.