Prefer the player's nickname over their username, and set accounts with unapproved names to their ID.
This ensures the correct name is shown everywhere, in games and mail etc.

We still use the username for logging, apart from Commands logs.

Note that setting the unapproved nickname to ID breaks the player from seeing their own nickname underneath their penguin. We probably need a better solution because of this.