Allow redemption codes to include awards that aren't clothing items
Awards now have a type property that defines whether it is clothing, furniture, an igloo, a location, a floor, a puffle or a puffle item.

We also include card as a type for the purposes of the card code types.

We also include an award ID which allows us to include multiple awards of the same ID, such as for giving more than one of a certain furniture item. This is not incremental.

Both award IDs and types are included as primary keys so that we can define different types with the same IDs (locations, floors, puffles all have low IDs) and the same item multiple times.

To keep things clean, we also now have a sendXt boolean for adding igloos, locations and floors.